% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibo.R
\name{fibo}
\alias{fibo}
\title{Calculates the nth fibonacci number}
\usage{
fibo(n)
}
\arguments{
\item{n}{}
}
\value{
the nth fibonacci number
}
\description{
Calculates the nth fibonacci number
}
\examples{
fibo(6)

fibo(5) == fibo(4) + fibo(3)
}
